John Denker: Add a view debugging functions and represent the viewer quats in the property tree for debugging.
Do a nontrivial calculation, demonstrating how to find the 12:00 orientation given the view and the view offset.  it, and observe that the 12:00 direction does not change when you change the view offset, which is a nontrivial result.

Erik: This already proved to be useful since I now obserbed that the sky and fog coloring code depends on view offset rather that the view orientations. This is why those effects are out of line in some views.