Bugfix: set temp and dewpoint from the gui
This fix (re)enables the setting of temperature and
dewpoint from the weather-conditions dialog and other
sources for various altitude layers. These temperatures
are reduced to the corresponding sea level temperatures
according to ICAO standard atmosphere which is the inverse
function of the calculation of temperature at altitude based
on sea-level-temperature.
Note: this only works for the troposphere.