Fix bug 182 - make sun direction agree with position of sun from the ephemeris
http://code.google.com/p/flightgear-bugs/issues/detail?id=182

There were two issues. The biggest is that the sunsolver was accessing
parameters from the ephemeris and doing a different calculation with
them to derive the right ascension and declination of the sun. I'm not
sure who is right, but I changed sunsolver to agree with the
ephemeris. Also, there was an inappropriate use of geodetic
coordinates in calculating  the sun latitude and longitude.

Also, I did some cleanup in updateSunPos().