ATIS improvements from John Denker
Primarily an improvment to the internationalisation.  Previously the ATIS considered
the world as either the UK or the rest of the world.  Now it considers it as US+Canada
or the rest of the world.  This is much more consistent with global usage of pressure
units as either inches or millibars, which is the most important difference.  Also
some other tweaks and cleanups.