replaced the outdated bash completion script with a completely rewritten version
The new script is documented in the flightgear wiki (see "bash completion") and adds functionality such as advanced detection of FG_ROOT and other paths, and many completions not available before.
The list of possible command line options is automatically generated from "fgfs --help --verbose".

The author of the original script (mfranz) has agreed to replace this script.