TestSuite: Converted the old tests/test-up.cxx program into a simgear unit test.
This old standalone test, from Curt back in 2001, was converted to the format
required for the CppUnit test suite.  As most of simgear/math/sg_geodesy.hxx no
longer exists, the code was updated using SGGeod and SGGeoc.  Hardcoded values
are used to permanently lock in the values currently calculated by simgear.