Security fix: don't allow the route manager to overwrite arbitrary files
Since the Save function can be triggered from Nasal with an arbitrary
path, we must check this path before overwriting the file.

(also add a missing include that is directly needed for this commit)

This commit is a backport of 280cd523686fbdb175d50417266d2487a8ce67d2,
from 'next'. It fixes CVE-2016-9956.